Add a demonstration of the linting in Drone CI
One of the more common tasks required by build systems is to express the
results of some sort of static analysis, be that lints, tests or so
fourth.

This commit adds a demonstration of adding Lints to Drone CI. These
lints are fairly simple, but are added and aggregated via the method
used in:

  https://bit.ly/2Btj6dU

== Design Notes

=== Composer

A common pattern within PHP applications is to include a set of "dev
dependencies"; modifications to the application runtime or libraries
that allow to be introspected or verified at a greater level.

In this case, the application "PHPCS" has been included in the
repository.

Initially the goal was to consume this PHPCS binary in the pipeline
also. However, this ran into a design constraint:

- The pipeline is mounted in a `noexec` environment
- Arcanist requires a 'binary' to execute lints

This is, in a sense, functioning as intended -- it is unusual to bundle
executable files in a build system. However, it means that the
application pattern commonly used in PHP cannot be used here.

Instead, we consume the external PHPCS in the pipeline.

=== Funny ref to master

Normally it would be reasonable to simply compare to "master" (aka
refs/heads/master). However, because this is a PR event and we want to
run tests etc. on the results of the merge Drone will merge the branch
into the destination before it run the analysis.

This process updates the local master.

However, we can still reference our "intended" master by referencing the
current state as it is on the origin:

  refs/remotes/origin/HEAD

This is an odd quirk of Git & the pipeline, but defensible.

=== `--ansi`

Reading the output of the terminal is an inherently "expensive"
exercise, and a much better integration would be with GitHubs "checks"
API. However, that is a little expensive for this work, requiring 1 - 2
days of development.

Because tasks are run in the pipeline they're run in a "non-interactive"
shell ANSI output is suppressed as its usually useless and may break
tools that parse the output. However, in this case we explicitly enable
it to display in the pipeline.